// src/store/auth/actions/types.ts
enum AuthActionTypes {
  CLEAR_AUTH_STATE = 'CLEAR_AUTH_STATE',
  /** Sign in / Login */
  SIGN_IN = 'SIGN_IN',

  SET_AUTH_INFO = 'SET_AUTH_INFO',
  CLEAR_AUTH_INFO = 'CLEAR_AUTH_INFO',
  SAVE_AUTH_INFO = 'SAVE_AUTH_INFO',
  LOAD_AUTH_INFO = 'LOAD_AUTH_INFO',

  SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS',

  FETCH_NEW_TOKEN = 'FETCH_NEW_TOKEN',

  /** Response */
  SET_SIGN_IN_RESPONSE = 'SET_SIGN_IN_RESPONSE',
  CLEAR_SIGN_IN_RESPONSE = 'CLEAR_SIGN_IN_RESPONSE',

  /** Sign out / Logout */
  SIGN_OUT = 'SIGN_OUT',
  SIGN_OUT_SUCCESS = 'SIGN_OUT_SUCCESS',

}

export default AuthActionTypes;